ref(metrics): Move MetricListItemDetails
