+++ static/app/components/metrics/mriSelect/metricListItemDetails.tsx
